Fix some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/9457



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@32111 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/regextst.cpp
@@ -309,7 +309,7 @@ void RegexTest::assertUTextInvariant(const char *expected, UText *actual, const 
 
 #define INV_BUFSIZ 2048 /* increase this if too small */
 
-static int32_t inv_next=0;
+static int64_t inv_next=0;
 
 #if U_CHARSET_FAMILY!=U_ASCII_FAMILY
 static char inv_buf[INV_BUFSIZ];
